requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Bleeding edge Django
django==2.0
# Configuration
django-environ==0.4.4
django-cors-headers==2.1.0


# Models
django-model-utils==3.0.0

# Images
Pillow==4.3.0

# Password storage
argon2-cffi==16.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.3.2

# Unicode slugification
awesome-slugify==1.6.5

pytz==2017.3

django-redis==4.8.0
redis>=2.10.5

celery==4.1.0


# Your custom requirements go here

python-telegram-bot==6.1.0
djangorestframework==3.7.7
django-celery-beat==1.1.0
